Updated vim. WIP on others
