+++ setups/i3/i3-config
@@ -12,9 +12,9 @@ bindsym $mod+Shift+Return exec terminator --role floating
 bindsym $mod+a exec pavucontrol
 
 # Rofi
-bindsym $mod+i exec "~/.local/bin/omni-menu run"
-bindsym $mod+o exec "~/.local/bin/omni-menu window"
-bindsym $mod+p exec "~/.local/bin/omni-menu"
+bindsym $mod+i exec "GSK_RENDERER=cairo ~/.local/bin/omni-menu run"
+bindsym $mod+o exec "GSK_RENDERER=cairo ~/.local/bin/omni-menu window"
+bindsym $mod+p exec "GSK_RENDERER=cairo ~/.local/bin/omni-menu >> /tmp/omni-menu.log 2>&1"
 
 # bindsym Control+space exec dunstctl close
 bindsym $mod+F1 exec zathura ~/docs/help.pdf
