--- setups/nvim/after/plugin/cmp.lua
@@ -1,58 +1,0 @@
-local cmp = require('cmp')
-local select_opts = { behavior = cmp.SelectBehavior.Insert }
-local lspkind = require('lspkind')
-
-cmp.setup({
-    snippet = {
-        expand = function(args)
-            vim.fn["vsnip#anonymous"](args.body) -- For `vsnip` users.
-        end,
-    },
-    window = {
-        -- completion = cmp.config.window.bordered(),
-        documentation = cmp.config.window.bordered(),
-    },
-    sources = cmp.config.sources({
-        { name = 'nvim_lsp', group_index = 2 },
-        { name = "copilot",  group_index = 2 },
-        { name = 'vsnip',    group_index = 2 },
-        { name = 'path',     group_index = 2 }
-    }, {
-        { name = 'buffer' },
-    }),
-    mapping = {
-        ['<C-k>'] = cmp.mapping.select_prev_item(select_opts),
-        ['<C-j>'] = cmp.mapping.select_next_item(select_opts),
-        ['<C-b>'] = cmp.mapping.scroll_docs(-4),
-        ['<C-f>'] = cmp.mapping.scroll_docs(4),
-        ['<CR>'] = cmp.mapping.confirm({ select = false }),
-        ['<C-Space>'] = cmp.mapping.complete(),
-    }
-})
-
--- Set configuration for specific filetype.
-cmp.setup.filetype('gitcommit', {
-    sources = cmp.config.sources({
-        { name = 'git' }, -- You can specify the `git` source if [you were installed it](https://github.com/petertriho/cmp-git).
-    }, {
-        { name = 'buffer' },
-    })
-})
-
--- Use buffer source for `/` and `?` (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline({ '/', '?' }, {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = {
-        { name = 'buffer' }
-    }
-})
-
--- Use cmdline & path source for ':' (if you enabled `native_menu`, this won't work anymore).
-cmp.setup.cmdline(':', {
-    mapping = cmp.mapping.preset.cmdline(),
-    sources = cmp.config.sources({
-        { name = 'path' }
-    }, {
-        { name = 'cmdline' }
-    })
-})
