--- setups/nvim/after/plugin/conform.lua
@@ -1,18 +1,0 @@
-require("conform").setup({
-    formatters_by_ft = {
-        lua = { "stylua" },
-        -- Conform will run multiple formatters sequentially
-        python = { "isort", "black" },
-        -- Use a sub-list to run only the first available formatter
-        javascript = { { "prettierd", "prettier" } },
-        typescript = { { "prettierd", "prettier" } },
-        javascriptreact = { { "prettierd", "prettier" } },
-        typescriptreact = { { "prettierd", "prettier" } },
-
-        json = { "jq" }
-    },
-    format_on_save = {
-        -- These options will be passed to conform.format()
-        lsp_fallback = true,
-    },
-})
