--- setups/nvim/after/plugin/folding.lua
@@ -1,44 +1,0 @@
--- vim.opt.foldmethod = "expr"
--- vim.opt.foldexpr = "nvim_treesitter#foldexpr()"
---
--- local vim = vim
--- local api = vim.api
--- local M = {}
--- -- function to create a list of commands and convert them to autocommands
--- -------- This function is taken from https://github.com/norcalli/nvim_utils
--- function M.nvim_create_augroups(definitions)
---     for group_name, definition in pairs(definitions) do
---         api.nvim_command('augroup ' .. group_name)
---         api.nvim_command('autocmd!')
---         for _, def in ipairs(definition) do
---             local command = table.concat(vim.tbl_flatten { 'autocmd', def }, ' ')
---             api.nvim_command(command)
---         end
---         api.nvim_command('augroup END')
---     end
--- end
---
--- local autoCommands = {
---     -- other autocommands
---     open_folds = {
---         { "BufReadPost,FileReadPost", "*", "normal zR" }
---     }
--- }
---
--- M.nvim_create_augroups(autoCommands)
---
-
-vim.o.foldcolumn = '1' -- '0' is not bad
-vim.o.foldlevel = 99   -- Using ufo provider need a large value, feel free to decrease the value
-vim.o.foldlevelstart = 99
-vim.o.foldenable = true
-
--- Using ufo provider need remap `zR` and `zM`. If Neovim is 0.6.1, remap yourself
-vim.keymap.set('n', 'zR', require('ufo').openAllFolds)
-vim.keymap.set('n', 'zM', require('ufo').closeAllFolds)
-
-require('ufo').setup({
-    provider_selector = function(bufnr, filetype, buftype)
-        return { 'treesitter', 'indent' }
-    end
-})
