--- setups/nvim/after/plugin/lean.lua
@@ -1,3 +1,0 @@
-require('lean').setup {
-    mappings = true,
-}
