--- setups/nvim/after/plugin/lint.lua
@@ -1,14 +1,0 @@
-local lint = require('lint')
-
-lint.linters_by_ft = {
-    javascript = { 'eslint_d' },
-    typescript = { 'eslint_d' },
-    javascriptreact = { 'eslint_d' },
-    typescriptreact = { 'eslint_d' }
-}
-
-vim.api.nvim_create_autocmd({ "BufWritePost" }, {
-    callback = function()
-        lint.try_lint()
-    end,
-})
