--- setups/nvim/after/plugin/lspkind.lua
@@ -1,8 +1,0 @@
-local lspkind = require("lspkind")
-lspkind.init({
-    symbol_map = {
-        Copilot = "",
-    },
-})
-
-vim.api.nvim_set_hl(0, "CmpItemKindCopilot", { fg = "#6CC644" })
