--- setups/nvim/after/plugin/neoscroll.lua
@@ -1,1 +1,0 @@
-require('neoscroll').setup()
