--- setups/nvim/after/plugin/startify.lua
@@ -1,1 +1,0 @@
-vim.g.startify_bookmarks = { { c = '~/.config/nvim' }, { o = "~/owl" }, { d = "~/docs" } }
