+++ setups/nvim/lua/plugins/completion.lua
@@ -1,0 +1,88 @@
+return {
+  -- Completion engine
+  {
+    'hrsh7th/nvim-cmp',
+    dependencies = {
+      'hrsh7th/cmp-nvim-lsp',
+      'hrsh7th/cmp-buffer',
+      'hrsh7th/cmp-path',
+      'hrsh7th/cmp-cmdline',
+      'hrsh7th/cmp-vsnip',
+      'hrsh7th/vim-vsnip',
+      'onsails/lspkind-nvim',
+    },
+    config = function()
+      local cmp = require('cmp')
+      local select_opts = { behavior = cmp.SelectBehavior.Insert }
+      local lspkind = require('lspkind')
+
+      cmp.setup({
+        snippet = {
+          expand = function(args)
+            vim.fn["vsnip#anonymous"](args.body) -- For `vsnip` users.
+          end,
+        },
+        window = {
+          documentation = cmp.config.window.bordered(),
+        },
+        sources = cmp.config.sources({
+          { name = 'nvim_lsp', group_index = 2 },
+          { name = "copilot",  group_index = 2 },
+          { name = 'vsnip',    group_index = 2 },
+          { name = 'path',     group_index = 2 }
+        }, {
+          { name = 'buffer' },
+        }),
+        mapping = {
+          ['<C-k>'] = cmp.mapping.select_prev_item(select_opts),
+          ['<C-j>'] = cmp.mapping.select_next_item(select_opts),
+          ['<C-b>'] = cmp.mapping.scroll_docs(-4),
+          ['<C-f>'] = cmp.mapping.scroll_docs(4),
+          ['<CR>'] = cmp.mapping.confirm({ select = false }),
+          ['<C-Space>'] = cmp.mapping.complete(),
+        }
+      })
+
+      -- Set configuration for specific filetype.
+      cmp.setup.filetype('gitcommit', {
+        sources = cmp.config.sources({
+          { name = 'git' },
+        }, {
+          { name = 'buffer' },
+        })
+      })
+
+      -- Use buffer source for `/` and `?`
+      cmp.setup.cmdline({ '/', '?' }, {
+        mapping = cmp.mapping.preset.cmdline(),
+        sources = {
+          { name = 'buffer' }
+        }
+      })
+
+      -- Use cmdline & path source for ':'
+      cmp.setup.cmdline(':', {
+        mapping = cmp.mapping.preset.cmdline(),
+        sources = cmp.config.sources({
+          { name = 'path' }
+        }, {
+          { name = 'cmdline' }
+        })
+      })
+    end,
+  },
+
+  -- LSP Kind
+  {
+    'onsails/lspkind-nvim',
+    config = function()
+      require('lspkind').init({
+        mode = 'symbol_text',
+        preset = 'codicons',
+        symbol_map = {
+          Copilot = "",
+        },
+      })
+    end,
+  },
+}
