+++ setups/nvim/lua/plugins/formatting.lua
@@ -1,0 +1,46 @@
+return {
+  -- Formatting
+  {
+    'stevearc/conform.nvim',
+    config = function()
+      require("conform").setup({
+        formatters_by_ft = {
+          lua = { "stylua" },
+          -- Conform will run multiple formatters sequentially
+          python = { "isort", "black" },
+          -- Use a sub-list to run only the first available formatter
+          javascript = { { "prettierd", "prettier" } },
+          typescript = { { "prettierd", "prettier" } },
+          javascriptreact = { { "prettierd", "prettier" } },
+          typescriptreact = { { "prettierd", "prettier" } },
+          json = { "jq" }
+        },
+        format_on_save = {
+          -- These options will be passed to conform.format()
+          lsp_fallback = true,
+        },
+      })
+    end,
+  },
+
+  -- Linting
+  {
+    'mfussenegger/nvim-lint',
+    config = function()
+      local lint = require('lint')
+
+      lint.linters_by_ft = {
+        javascript = { 'eslint_d' },
+        typescript = { 'eslint_d' },
+        javascriptreact = { 'eslint_d' },
+        typescriptreact = { 'eslint_d' }
+      }
+
+      vim.api.nvim_create_autocmd({ "BufWritePost" }, {
+        callback = function()
+          lint.try_lint()
+        end,
+      })
+    end,
+  },
+}
