+++ setups/nvim/lua/plugins/utilities.lua
@@ -1,0 +1,68 @@
+return {
+  -- Folding
+  {
+    'kevinhwang91/nvim-ufo',
+    dependencies = { 'kevinhwang91/promise-async' },
+    config = function()
+      vim.o.foldcolumn = '1' -- '0' is not bad
+      vim.o.foldlevel = 99   -- Using ufo provider need a large value, feel free to decrease the value
+      vim.o.foldlevelstart = 99
+      vim.o.foldenable = true
+
+      -- Using ufo provider need remap `zR` and `zM`. If Neovim is 0.6.1, remap yourself
+      vim.keymap.set('n', 'zR', require('ufo').openAllFolds)
+      vim.keymap.set('n', 'zM', require('ufo').closeAllFolds)
+
+      require('ufo').setup({
+        provider_selector = function(bufnr, filetype, buftype)
+          return { 'treesitter', 'indent' }
+        end
+      })
+    end,
+  },
+
+  -- Comments
+  {
+    'numToStr/Comment.nvim',
+    config = function()
+      require('Comment').setup()
+    end
+  },
+
+  -- Surround
+  {
+    "tpope/vim-surround",
+  },
+
+  -- Git integration
+  {
+    "tpope/vim-fugitive",
+  },
+
+  -- Markdown preview
+  {
+    "iamcco/markdown-preview.nvim",
+    build = function() vim.fn["mkdp#util#install"]() end,
+  },
+
+  -- Lean theorem prover
+  {
+    "Julian/lean.nvim",
+    config = function()
+      require('lean').setup({
+        abbreviations = { builtin = true },
+        lsp = { on_attach = on_attach },
+        lsp3 = false,
+        mappings = true,
+      })
+    end,
+  },
+
+  -- Smooth scrolling
+  {
+    'karb94/neoscroll.nvim',
+    config = function()
+      require('neoscroll').setup()
+    end,
+  },
+}
