+++ setups/tmux/tmux.conf
@@ -4,19 +4,11 @@ setw -g mode-keys vi
 
 set-option -g status-position top
 
-
-# List of plugins
+# plugins
 set -g @plugin 'tmux-plugins/tmux-yank'
-
 set -g @plugin 'dracula/tmux'
 set -g @dracula-plugins "git"
 set -g @dracula-show-left-icon session
-
-# Other examples:
-# set -g @plugin 'github_username/plugin_name'
-# set -g @plugin 'github_username/plugin_name#branch'
-# set -g @plugin 'git@github.com:user/plugin'
-# set -g @plugin 'git@bitbucket.com:user/plugin'
 
 # Initialize TMUX plugin manager (keep this line at the very bottom of tmux.conf)
 run '~/.local/share/tmux/plugins/tpm/tpm'
